Move the Zuora code into its' own class and refactor to use that.
fix
